tests/auto_causal/methods/difference_in_differences/test_did_diagnostics.py
import pytest
import pandas as pd
import numpy as np
from auto_causal.methods.difference_in_differences.diagnostics import validate_parallel_trends

# Fixture (can reuse from estimator tests if needed, or define simpler one)
@pytest.fixture
def sample_did_data_diag():
    df = pd.DataFrame({
        'time': [1, 2, 3, 1, 2, 3],
        'unit': ['A', 'A', 'A', 'B', 'B', 'B'],
        'outcome': [10, 11, 12, 15, 17, 19],
        'group': [0, 0, 0, 1, 1, 1] # A is control, B is treated
    })
    return df

# Test Cases
def test_validate_parallel_trends_placeholder(sample_did_data_diag):
    """Tests the placeholder parallel trends validation function."""
    # For placeholder, specific args don't matter much
    results = validate_parallel_trends(
        sample_did_data_diag, 
        time_var='time', 
        outcome='outcome', 
        group_indicator_col='group',
        treatment_period_start=3, # Example treatment start
        dataset_description=None # Add arg
    )
    
    assert isinstance(results, dict)
    # Check for specific placeholder values if they are defined
    assert results.get('valid') is True # Function defaults to True when test cannot be run
    # Check the actual detail message returned when test fails on this data
    assert "Insufficient pre-treatment data or variation" in results.get('details', "")

# Add tests here if/when parallel trends validation is implemented
# e.g., test_parallel_trends_pass, test_parallel_trends_fail 