requirement.txt
pandas
vertexai
openai
google-cloud-aiplatform
docker
jupyter_client
ipykernel
pyzmq
requests
python-dotenv
statsmodels==0.14.4
seaborn==0.13.2
shap==0.43.0
shapely==2.0.7
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
regex==2024.11.6
requests==2.32.3
langchain-anthropic==0.3.10
langchain-core==0.3.50
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langchain-together==0.1.7
langsmith==0.3.24
lightgbm==4.6.0
llvmlite==0.44.0
markdown-it-py==3.0.0
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
dowhy==0.12
econml==0.15.1


